Add I777 NFC files - this allows most NFC functionality currently present in the I777 to work on Hellraised ROMs.
Some ROMs may need Settings.apk edits to enable the NFC entry - but on
VillainROM 3.0 (XXKI3), no mods are needed.

Note that I777 ROMs set the protection level for the STOP_APP_SWITCHES
permission to 3 (signed and system), while I9100 ROMs set it to 2.
This causes tapping of NDEF tags to force close the NFC service.